@@ -1,15 +1,21 @@
-[tool.poetry]
+[project]
 name = "taskiq-psqlpy"
 version = "0.1.8"
 description = "PSQLPy and PostgreSQL integration for taskiq"
-authors = ["taskiq-team <taskiq@norely.com>"]
+authors = [
+    { name = "Taskiq team", email = "taskiq@no-reply.com" }
+]
+maintainers = [
+    { name = "Taskiq team", email = "taskiq@no-reply.com" }
+]
+license = "MIT"
+license-files = ["LICENSE"]
 readme = "README.md"
 packages = [{ include = "taskiq_psqlpy" }]
 classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -26,25 +32,34 @@ keywords = [
     "result_backend",
     "psqlpy",
 ]
+requires-python = ">=3.10"
+dependencies = [
+    "psqlpy>=0.11.11",
+    "taskiq>=0.12.0",
+]
 
-[tool.poetry.dependencies]
-python = "^3.9.0"
-psqlpy = "^0.11.1"
-taskiq = "^0.11.17"
-
-[tool.poetry.group.dev.dependencies]
-black = "^23.1.0"
-pre-commit = "^2.20.0"
-mypy = "^1.16"
-flake8 = "^6"
-autoflake = "^1.4"
-yesqa = "^1.4.0"
-pytest = "^7.2.2"
-pytest-xdist = "^3.2.1"
-anyio = "^3.6.2"
-pytest-cov = "^4.0.0"
-wemake-python-styleguide = "^0.18.0"
-ruff = "^0.12"
+[dependency-groups]
+dev = [
+    {include-group = "lint"},
+    {include-group = "test"},
+    "pre-commit>=4.5.0",
+    "anyio>=4.12.0",
+]
+test = [
+    "pytest>=9.0.1",
+    "pytest-cov>=7.0.0",
+    "pytest-env>=1.2.0",
+    "pytest-xdist>=3.8.0",
+]
+lint = [
+    "black>=25.11.0",
+    "mypy>=1.19.0",
+    "ruff>=0.14.7",
+    "autoflake>=2.3.1",
+    "wemake-python-styleguide>=1.4.0",
+    "flake8>=7.3.0",
+    "yesqa>=1.5.0",
+]
 
 
 [tool.mypy]
@@ -59,8 +74,12 @@ allow_untyped_decorators = true
 warn_return_any = false
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["uv_build>=0.9.13,<0.10.0"]
+build-backend = "uv_build"
+
+[tool.uv.build-backend]
+module-root = ""
+module-name = "taskiq_psqlpy"
 
 [tool.ruff]
 line-length = 88
@@ -94,6 +113,8 @@ select = [
     "ERA", # Checks for commented out code
     "PL", # PyLint checks
     "RUF", # Specific to Ruff checks
+    "FA102", # Future annotations
+    "UP", # Pyupgrade
 ]
 ignore = [
     "D105", # Missing docstring in magic method
